Extract the crypto lowering group into a host submodule
Fourth cranelift_backend.rs split under #1254, and the second of the
direct-native i64 lowering by capability group (after host_fs). A
call-closure analysis seeded on the crypto family identified 24 i64
lowering functions (sha256/HMAC/constant-time-eq audited-condition
helpers, random-bytes lowering, and the compile-time spike_crypto_*
AEAD/Ed25519 dispatchers) with exactly one callback into the recursive
expr/stmt lowering hub (lower_i64_expr).

Move those 24 functions plus the crypto-only OpenSSL FFI machinery they
exclusively depend on (SpikeEd25519Crypto/SpikeAeadCrypto Drop impls, the
SpikeEd25519PkeyCtxGuard/PkeyGuard/MdCtxGuard/SpikeAeadCtxGuard RAII
guards and their Drop impls, the SPIKE_EVP_* constants, and the
dlopen/dlsym/dlclose symbol-loading helpers) into
stage1/crates/axiomc/src/cranelift_backend/host_crypto.rs.
cranelift_backend.rs drops 22,728 -> 21,948 lines and the top-file line
share falls to 0.7581.

Pure code motion. The shared crypto_* wrapper functions the compile-time
evaluator calls (crypto_random_bytes, crypto_ed25519_keygen, etc.) stay in
cranelift_backend/intrinsics.rs; they and the new host_crypto.rs sibling
reach the crypto-loading helpers moved here through the parent's existing
`pub(crate) use host_crypto::*` re-export, the same pattern host_fs.rs
established. I64HelperSignature is raised to pub(crate) alongside the
already-pub(crate) I64StaticBindings to keep the moved signatures visible.

Refs #1254.
